+++ user/src/main/java/org/group21/user/model/User.java
@@ -13,7 +13,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "first_name", nullable = false)
     private String firstName;
@@ -24,6 +24,9 @@ public class User {
     @Column(nullable = false, unique = true)
     private String email;
 
+    // For simplicity, the password is stored as plain text
+    // However, in a real-life scenario, this would be stored as a salted hash
+    // to avoid storing plain text passwords
     @Column(nullable = false)
     private String password;
